[IMP] sale_timesheet: allow to timesheet on project without putting the task_id
Currently timesheeting on a project without specifying the task does not
increase the delivered_qty of the related sale order line.

This could potentially lead to issues with the Helpdesk module for instance.
When a project is allow_billable = false, the user is not required to specify
the task_id on the helpdesk's ticket. This means that hours timesheeted while
working on this ticket will not be added to the delivered_qty of the related
SOL.

This commit "fixes" this by allowing to timesheet on project without specifying
the task_id. So now, even if an helpdesk's ticket has no task_id, the timesheeted
hours will be added to the related SOL delivered_qty.

Because it was initially impossible to timesheet on project without setting
the task_id, some tests had to be rewritten.

closes odoo/odoo#53172

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>