[FIX] point_of_sale: check payment methods intermediary accounts is set
This is normally not supposed to happen to have a payment method without a receivable account set,
as this is a required field. However, it happens the receivable account cannot be found during upgrades
and this is a bommer to block the upgrade for that point, given the user can correct this by himself,
without requiring a manual intervention from our upgrade support.
However, this must be ensured this receivable is well set before opening a POS session.

closes odoo/odoo#52862

X-original-commit: 838c0d86067c04d25cbdb8c6a859e23572e67339
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>