[FIX] website: hide countdown at the end
When using the options 'Hide countdown at the end', only the first
canvas would dissapear and the message wasn't showing up. It was due
to a return in a loop that would cause the code to stop at that point
instead of just going to the next iteration as intented.

task-2261649

closes odoo/odoo#53218

X-original-commit: 1dda65809a61d3acd98791f6aa6a64cec356b6eb
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>