[FIX] web_editor: consider ]0, 1] values as 'active' values in editor
Before this commit, the left panel input widgets were considering the
values between 0 and 1 as 'inactive' and thus were forcing their
default value if the user entered those. As the default is most of the
time 0, entering 0.9 gave a 0 result for no apparent reason.

Part of https://github.com/odoo/odoo/pull/52821

closes odoo/odoo#52841

X-original-commit: 069084bf919d31c8595f650499c9ce10e18caf0b
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>