[IMP] sms/mass_mailing: add emojis support to the base SmsWidget
This commit adds emojis to base SmsWidget in order to use a text field with
both sms and emoji widgets in "mass_mailing_sms" module for sms content.

The emoji feature is optional (depends on "enable_emojis" attribute). To add
the emoji support, the following changes have been made:

  * override the method "on_attach_callback" from emoji widget base code to
    hide emojis if not enabled;

We also adjust the css of emoji widget to make it suitable for a text field
with sms option.

Task ID-2224393
PR #49096