[FIX] web_editor: fix traceback when clicking on website logo in edit
When clicking on an image in edit mode, the web editor attempts to find
an attachment that corresponds to the image's URL. In the case of the
website logo, the URL looks like it's referencing an attachment by id or
xml-id, but it's actually neither. This cause a variable to never be
initialized, which in turns causes a traceback when trying to access it
later.

This commit fixes that by always initializing attachment to None.

closes odoo/odoo#53418

X-original-commit: 5d962b8a0a238e106489e7975bd0bf0c63402c1f
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>