[FIX] website_event: stop displaying messages on event page
A recent refactoring of mail.message induced a crash when redering the
front side of an event with messages/notes in its chatter. As there is
no use to such a feature, we simply remove the part that displays those
messages.

Task ID 2279183
Linked to task ID 2071556 (message publish management cleaning)

closes odoo/odoo#53135

X-original-commit: d2df5c73c285e4efb89d2f6d7d9d4c3e17a170de
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>