[FIX] mrp: create backorder with no components
Consuming all the components for only a part of the finished product
will ask the manufacturing user to create a backorder. Confirm it will
fail as there are no component left to consume to produce the
remaining quantity of finished product. This commit only confirm
backorders that still have some quantities to consume.

Task : 2278147

X-original-commit: cc2d060a254e384001ee7ca55a2041e75579fe3c