[FIX] account: sequence fetch next sequence correctly
Before, if we had PREFIX/9 and PREFIX/10, the highest number was the
lexicographic order, leading to proposing PREFIX/10 again and again.

Because of performances, we can't execute the regex in the SELECT query
and we need to store it in the database. This will allow to do an
indexed search.

closes odoo/odoo#54811

X-original-commit: d9fad36ffc81f6d353993c53f8cc9df3d99bdc37
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>