[FIX] crm : prevent a false positive warning ribbon for phone on lead
In the lead, we compute the phone number to make it formatted to match
with the international phone number format.

Before this commit, this formatted number was used for comparison with
partner's phone (which is not formatted), and so, even if both the
numbers are same, the warning ribbon appears due to different formats
which is not correct behavior.

This commit fixes the issue by comparing both the formatted number to
remove the false positive warning ribbon and thus making sure that it
appears only when there actually is a difference between lead's phone
and the phone number set on partner.

Task Id-2274752

closes odoo/odoo#54942

X-original-commit: df524cc7e2ee21bbd91c008bb835ba833402e9b5
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>